8244732: Shenandoah: move heuristics code to gc/shenandoah/heuristics Reviewed-by: rkennke
