+++ src/hotspot/share/gc/shenandoah/shenandoahPhaseTimings.cpp
@@ -27,8 +27,8 @@
 #include "gc/shenandoah/shenandoahCollectorPolicy.hpp"
 #include "gc/shenandoah/shenandoahPhaseTimings.hpp"
 #include "gc/shenandoah/shenandoahHeap.inline.hpp"
-#include "gc/shenandoah/shenandoahHeuristics.hpp"
 #include "gc/shenandoah/shenandoahUtils.hpp"
+#include "gc/shenandoah/heuristics/shenandoahHeuristics.hpp"
 #include "runtime/orderAccess.hpp"
 #include "utilities/ostream.hpp"
